# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OllO API
  version: 1.0.0
  description: |
    Store-and-forward API for sealed envelopes. The server does not parse
    message plaintext. See docs/api/API.md.
servers:
  - url: /v1
paths:
...
  /me:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Current user }
    put:
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Updated }
  /envelopes:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200": { description: Mailbox drain }
    post:
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Accepted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer